5/4 Changes
Note 1 - Trying to get the main class to run the validator. Created object named validate and tried 
to call isValid() from the InetAddressValidator class. It skipped right over it and ran the greeting
and test. I believe the Inet class is missing something.

Note 2: Downloaded the files and added the JAR files to this project in Netbeans